+++ test/unit/fuzz-tools.test.js
@@ -168,6 +168,7 @@ import * as ltcgaV173 from '../../lib/ltcga-v173.js';
 import * as ltcgaV174 from '../../lib/ltcga-v174.js';
 import * as ltcgaV175 from '../../lib/ltcga-v175.js';
 import * as ltcgaV176 from '../../lib/ltcga-v176.js';
+import * as ltcgaV177 from '../../lib/ltcga-v177.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -280,6 +281,7 @@ const MODULES = {
   'ltcga-v174.js': ltcgaV174,
   'ltcga-v175.js': ltcgaV175,
   'ltcga-v176.js': ltcgaV176,
+  'ltcga-v177.js': ltcgaV177,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];